requirements_frozen.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
Cython==0.29.37
frozenlist==1.4.0
h11==0.16.0
idna==3.11
multidict==6.0.4
pyenet==1.3.17
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.1.0
wsproto==1.3.2
yarl==1.9.3
